Data extraction module contenarized
