+++ data_extraction/Dockerfile
@@ -1,0 +1,9 @@
+FROM python:latest
+
+WORKDIR /app
+
+COPY . /app/
+
+RUN pip install --trusted-host pypi.python.org -r requeriments.txt
+
+ENTRYPOINT [ "python", "web.py" ]
