+++ data_extraction/last_extraction_date.txt
@@ -1,1 +1,1 @@
-05/06/2024
+24/06/2024
