+++ data_extraction/requeriments.txt
@@ -1,0 +1,5 @@
+selenium==4.9.1
+beautifulsoup4==4.12.2
+requests
+lxml
+pandas
