+++ docker-compose.yml
@@ -1,0 +1,29 @@
+version: '3'
+
+services:
+  selenium:
+    image: selenium/standalone-chrome:latest
+    container_name: selenium_service
+    shm_size: 2g
+    ports:
+      - "4444:4444"
+    networks:
+      - selenium_net
+
+  data_extraction:
+    build:
+      context: ./data_extraction
+      dockerfile: Dockerfile
+    ports:
+      - "4040:4040"
+    volumes:
+      - ./data_extraction:/app
+    depends_on:
+      - selenium
+    networks:
+      - selenium_net
+  
+networks:
+  selenium_net:
+    driver: bridge
+
